requirements.txt
Django
django-cors-headers
djangorestframework
python-decouple
whitenoise
pillow
torch
torch-vision